+++ packages/gatsby-source-prismic/src/types.ts
@@ -75,6 +75,7 @@ export interface Dependencies {
 	reporter: gatsby.Reporter;
 	reportInfo: gatsby.Reporter["info"];
 	reportWarning: gatsby.Reporter["warn"];
+	reportVerbose: gatsby.Reporter["verbose"];
 	globalNodeHelpers: NodeHelpers;
 	nodeHelpers: NodeHelpers;
 	pluginOptions: PluginOptions;