+++ packages/gatsby-source-prismic/test/__testutils__/createGatsbyContext.ts
@@ -39,6 +39,7 @@ export const createGatsbyContext = (): PartialDeep<gatsby.NodePluginArgs> & {
 			createFieldExtension: sinon.stub(),
 		},
 		reporter: {
+			verbose: sinon.stub(),
 			info: sinon.stub(),
 			warn: sinon.stub(),
 		},